Server Navmesh 오류 수정을 위한 최적 정리 커밋   - Unity에서 navmesh Export시 Test에 맞게 파일명 변경, 이름 변경, Lua 제작 주석 처리   - NavmeshTest를 libdetour 오류 해결을 위한 기본 CellSize수정
현제 문제점 :
  - Navmesh를 이용해 특정 위치까지는 정상적으로 이동이 가능하다.
  - 특정 위치 혹은 다리를 건널수가 없게 되어있다.

해결 방안 :
  - 기존 Navmesh 원본과 다른점을 찾아서 해결 하여야 할듯하다.
  - Navmesh 움직이는 최대 거리가 있는지 확인해야 할듯하다.
  - 모르겠다...

+++ Server/Navmesh_Ctrl/libdetour/dllexport.cpp
@@ -41,7 +41,7 @@ extern "C" {
 
 	void dh_get_triangles(const void* ug, void* &start, int* length) {
 		auto obj = (DetourHelper*)ug;
-		*length = obj->GetTriangles().size();
+		*length = (int)obj->GetTriangles().size();
 		start = (void*)obj->GetTriangles().data();
 #if _DEBUG
 		cout << "triangles:" << *length << endl;
@@ -50,7 +50,7 @@ extern "C" {
 	}
 	void dh_get_vectices(const void* ug, void* &start, int* length) {
 		auto obj = (DetourHelper*)ug;
-		*length = obj->GetVertices().size();
+		*length = (int)obj->GetVertices().size();
 		start = (void*)obj->GetVertices().data();
 #if _DEBUG
 		cout << "dh_get_vectices:" << *length << "\tstart:" << start << endl;
@@ -58,7 +58,7 @@ extern "C" {
 	}
 	void dh_get_indices(const void* ug, void* &start, int* length) {
 		auto obj = (DetourHelper*)ug;
-		*length = obj->GetIndices().size();
+		*length = (int)obj->GetIndices().size();
 		start = (void*)obj->GetIndices().data();
 #if _DEBUG
 		cout << "dh_get_indices:" << *length << "\tstart:" << start << endl;
@@ -72,7 +72,7 @@ extern "C" {
 		DHVertex v1 = *(DHVertex*)start;
 		DHVertex v2 = *(DHVertex*)end;
 		bool success = obj->FindPath(v1, v2, *path);
-		*length = path->size();
+		*length = (int)path->size();
 #if _DEBUG
 		cout << "dh_findpath: size:" << *length << endl;
 #endif
@@ -129,7 +129,7 @@ extern "C" {
 		DHVertex v1 = *(DHVertex*)start;
 		DHVertex v2 = *(DHVertex*)end;
 		bool success = obj->FindStraightPath(v1, v2, *path);
-		*length = path->size();
+		*length = (int)path->size();
 #if _DEBUG
 		cout << "dh_find_straight_path: size:" << *length << endl;
 #endif
